% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier.R
\name{bezier}
\alias{bezier}
\title{Create Bezier-curves using the Bernstein approximation.}
\usage{
bezier(x = c(0, 1, -1), y = c(0, 1.5, 1.5), w = 1, resol = 100)
}
\arguments{
\item{x}{a numeric vector giving the x-coordinates of exactly 
three Bezier-points. Defaults to c(0,1,-1) matching the 'mt'
format in \link{mt_align}.}

\item{y}{a numeric vector giving the x-coordinates of exactly
three Bezier-points. Defaults to c(0,1.5,1.5) matching the 'mt'
format in \link{mt_align}.}

\item{w}{a numeric giving one more Bezier coefficients that, which 
controls the pull towards the middle point. Each entry in 
\code{w} creates one Bezier-curve.}
}
\description{
\code{mt_bezier} creates 3-point Bezier-curves using the 
  Bernstein appproximation to simulate continuous competition 
  in mouse- and hand-trajectories.
}
\examples{
\dontrun{
  # Generate range of Bezier-curves
  b = bezier(w=seq(0,10,.1))
  plot.new();plot.window(c(-1,1),c(0,1.5))
  for(i in 1:dim(b)[1]) lines(b[i,,])
}
   
}
\author{
Dirk U. Wulff (\email{dirk.wulff@gmail.com})
}
