% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_helpers.R
\name{fix_ids}
\alias{fix_ids}
\title{Create or replace ID codes}
\usage{
fix_ids(data)
}
\arguments{
\item{data}{A datset}
}
\value{
a tibble or a dataframe with fixed IDs
}
\description{
Create or replace ID codes
}
\examples{
 \dontrun{
 fix_ids(sniffy_data)
}
}
