% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausswindow.R
\name{gausswindow}
\alias{gausswindow}
\title{Create a Gausswindow.}
\usage{
gausswindow(l, w = 2.5, ...)
}
\arguments{
\item{l}{length of the window.}

\item{w}{width of the gauss curve in the window.}

\item{...}{ignored.}
}
\value{
A numeric vector of length \code{l}.
}
\description{
Create a Gausswindow of specified length and width as a numeric vector.
This function can be used for smoothing of trajectory data. The gausswindow
can be applied to a numeric vector using the convolve function.
}
\examples{
# Create some example data, just some arbitrary sum of sines
# NOTE: smoothed version will be a sum of sines of same
#       frequency as sines and cosines are eigenvectors
#       of convolution.
x <- seq(-10,10,by=0.1)
y <- 5*sin(2*x*2*pi+0.5)+2*sin(3*x*2*pi+0.3)+1*sin(4*x*2*pi+0.1)

\dontrun{
   plot(x,y,type="l")
}

# create a Gausswindow of length 10
g <- gausswindow(10)

# Convolve y with g to create a smoothed version
# NOTE: s will be shorter than y, if type="filter"
#       and longer if type="open"
s <- convolve(y,g,type="filter")

\dontrun{
   plot(x[seq_along(s)],s,type="l")
}

}
\author{
Tillmann Nett
}
