% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Sniffy data for the take-away paper}
\usage{
get_data(cand_no)
}
\arguments{
\item{cand_no}{Candidate number}
}
\value{
A tibble of Sniffy data.
}
\description{
Get Sniffy data for the take-away paper
}
\examples{
get_data(123456)


}
