% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Koop2013}
\alias{Koop2013}
\title{Koop & Johnson (2013) dataset}
\format{An object of class \code{data.frame} with 256035 rows and 5 columns.}
\usage{
Koop2013
}
\description{
Koop & Johnson (2013) dataset
}
\keyword{datasets}

