% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_helpers.R
\name{make_ids}
\alias{make_ids}
\title{Make random ID codes}
\usage{
make_ids(n)
}
\arguments{
\item{n}{Number of characters}
}
\value{
a character verctor of IDs
}
\description{
Make random ID codes
}
\examples{
 \dontrun{
 make_ids(10)
}
}
