% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_helpers.R
\name{make_sniffy_data}
\alias{make_sniffy_data}
\title{Create random Sniffy-style data}
\usage{
make_sniffy_data(cand_no, n = 1000, messy = TRUE)
}
\arguments{
\item{cand_no}{A candidate number}

\item{n}{Number of observations to create per condition}

\item{messy}{Should the output be "messed up" with bad values? Default is TRUE.}
}
\value{
A tibble of Sniffy data
}
\description{
Create random Sniffy-style data
}
\examples{
 \dontrun{
 make_sniffy_data(12345)
}

}
