% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mt_add_variables}
\alias{mt_add_variables}
\title{Add new variables to trajectory array.}
\usage{
mt_add_variables(data, use = "trajectories", save_as = use, variables)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import 
functions (see \link{mt_example} for details). Alternatively, a trajectory 
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be 
used.}

\item{save_as}{a character string specifying where the resulting trajectory 
data should be stored.}

\item{variables}{a character vector specifying the name of the new variables
that should be added to the trajectory array.}
}
\value{
A mousetrap data object (see \link{mt_example}) where the new 
  variables have been added as additional columns to the trajectory array. 
  The values for the added variables are \code{NA} and have to be filled in 
  manually (see Examples). If columns of the same name already existed, they 
  have been removed. If the trajectory array was provided directly as 
  \code{data}, only the trajectory array will be returned.
}
\description{
Add new variables to the trajectory array (and remove potentially existing 
variables of the same name). This is mostly a helper function used by other 
functions in this package (e.g., \link{mt_calculate_deviations}). However, it
can also be helpful if the user has calculated new variables for each logged 
coordinate and wants to add them to an existing trajectory array.
}
\examples{
# Add a new variable (called "xy_sum") to the trajectory array
mt_example <- mt_add_variables(mt_example,variables="xy_sum")

# Fill the new variable with the corresponding values
# (in this arbitrary example the sum of the x- and y-positions)
mt_example$trajectories[,"xy_sum",] <- 
  mt_example$trajectories[,"xpos",] + mt_example$trajectories[,"ypos",]


}

