% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mt_align}
\alias{mt_align}
\title{Align trajectories}
\usage{
mt_align(data, use = "trajectories", dimensions = c("xpos", "ypos"),
  save_as = "al_trajectories", start = TRUE, end = TRUE,
  coordinates = "norm")
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import 
functions (see \link{mt_example} for details). Alternatively, a trajectory 
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be 
used.}

\item{dimensions}{a character string specifying which trajectory variables 
should be used. Can be of length 2 or 3 for 2-dimensional or 3-dimensional
alignment respectively.}

\item{save_as}{a character string specifying where the resulting trajectory 
data should be stored.}

\item{start}{boolean specifying wheter the trajectorie's start points should 
be aligned.}

\item{end}{boolean specifying wheter the trajectorie's end points should 
be aligned.}

\item{end}{either a numeric vector specying the xstart, ystart, xend, yend
coordinates of the trajectory start and end points, or 'mt' to set the 
coordinates to c(0,0,-1,1.5) or 'norm' to set coordinates to c(0,0,1,1). 
In the 3-dimensional case use coordinates is a vector of size 6.}
}
\value{
A mousetrap data object (see \link{mt_example}) with an additional
  array (by default called \code{aligned_trajectories}) containing the 
  aligned trajectories. If a trajectory array was provided directly as
  \code{data}, only the aligned trajectories will be returned.
}
\description{
\code{trajectory_align} aligns trajectories to common start point, end point, 
  and coordinate system.
}
\examples{
mt_example <- trajectory_align(mt_example,
  save_as="aligned_trajectories",  dimensions = c('xpos','ypos'),
  start = TRUE,end = TRUE,coordinates = 'mt')

}
\author{
Dirk U. Wulff
}

