% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mt_average}
\alias{mt_average}
\title{Average trajectories across constant time intervals.}
\usage{
mt_average(data, use = "trajectories", save_as = "av_trajectories",
  interval_size = 100, max_time = NULL, show_progress = TRUE)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import 
functions (see \link{mt_example} for details).}

\item{use}{a character string specifying which trajectory data should be 
used.}

\item{save_as}{a character string specifying where the resulting trajectory 
data should be stored.}

\item{interval_size}{an integer specifying the size of the constant time 
interval. The unit corresponds to the unit of the timestamps.}

\item{max_time}{an integer specifying the upper time limit of the last time 
interval that should be included (therefore, it should be a multiple of the
\code{interval_size}). If specified, only values will be used for averaging
where the timestamps are smaller than \code{max_time}. If unspecified (the
default), all values will be included.}

\item{show_progress}{logical indicating whether function should report its 
progress.}
}
\value{
A mousetrap data object (see \link{mt_example}) with an additional
  array (by default called \code{av_trajectories}) that contains the average
  trajectory data per time interval.
  
  For the timestamps, the mid point of the respective time interval is 
  reported, which is helpful for plotting the trajectory data later on. 
  However, this value does not necessarily correspond to the empirical mean 
  of the timestamps in the interval.
}
\description{
Average trajectory data across constant time intervals. For every variable 
that was recorded and stored in the trajectory array (x- and y-position, 
possibly also velocity and acceleration etc.), the mean value for the 
respective time interval is calculated (see Details for information regarding
the exact averaging procedure).
}
\details{
For each time interval, it is first determined which of the timestamps lie 
within the respective time interval. Intervals are left-open, right-closed 
(e.g., a timestamp of 1200 would be included in the interval 1100-1200 while 
a timestamp of 1300 would be included in the interval 1200-1300). Then, all 
values for which the corresponding timestamps lie within the interval are 
averaged.

In case the last interval is not fully covered by the timestamps (e.g., if 
the last timestamp has the value 1250), values for the corresponding interval
(1200-1300) will be computed based on the average of the values up to the 
last timestamp.

If average velocity and acceleration are of interest, 
\link{mt_calculate_derivatives} should be called before averaging.
}
\examples{
mt_example <- mt_calculate_derivatives(mt_example)
mt_example <- mt_average(mt_example, save_as="av_trajectories",
  interval_size=100)

}
\seealso{
\link{mt_calculate_derivatives} for calculating velocity and 
acceleration.

\link{mt_downsample} for resampling trajectories using a constant time 
interval.
}

