% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{mt_bind}
\alias{mt_bind}
\title{Join two trajectory arrays}
\usage{
mt_bind(trajectories1, trajectories2, verbose = FALSE)
}
\arguments{
\item{trajectories1}{a trajectory array (see \link{mt_example}).}

\item{trajectories2}{a trajectory array (see \link{mt_example}).}

\item{verbose}{logical indicating whether function should report its
progress.}
}
\value{
A trajectory array.
}
\description{
Join two trajectory arrays.
}
\examples{

\dontrun{
trajectories_combined <- mt_bind(
  mt_example$trajectories,
  mt_prototypes
)
}
}

