% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{mt_clean}
\alias{mt_clean}
\title{Clean trajectories}
\usage{
mt_clean(data, use = "trajectories", dimensions = c("xpos", "ypos"),
  save_as = "measures", method = "intersect", box_limits = c(-2, 2, -0.2,
  2), sim_limit = 2, minkowski_p = 2, clean = TRUE)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import 
functions (see \link{mt_example} for details). Alternatively, a trajectory 
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory object should be
used.}

\item{dimensions}{a character string specifying which trajectory variables 
should be used. Can be of length 2 or 3 for 2-dimensional or 3-dimensional
trajectories.}

\item{save_as}{a character string specifying how the clean variables should be
stored. If save_as is an appropriately-sized data frame existing in data, 
the variables are added, otherwise a new data frame will be created.}

\item{box_limits}{a numeric vector matching the length of dimensions that 
specifying the maximum allowed distance from the start point (measured
in terms of the distance between the trajectorie's start and end point).}

\item{sim_limit}{a numeric specifying the maximum allowed z-value
of a trajectory relative to the distribution of average trajectory 
similarities.}

\item{minkowski_p}{numeric specifying the distance function used to
assess similarity between trajectories. \code{minkowski_p} of 1 and 2 
translate into city-block and euclidian distances.}

\item{clean}{character specifying the variable used to clean the data. Can
be "box", "sim", "union, and "intersection", which will remove trials 
identified by the box_limit assessment, the similarity_limit, one of them,
or both, respectively.}

\item{aligned_sim}{boolean specifying whether the trajectories specified
in \code{use} should be aligned (see \link{trajectory_align} for details) 
or not before assessing similarity between trajectorie.}

\item{rescaled_sim}{boolean specifying whether the trajectories specified
in \code{use} should be spatially normalized (see \link{spatial_rescale} 
for details) or not before assessing similarity between trajectories. If 
FALSE \code{n_rescale} is ignored.}
}
\value{
A mousetrap data object (see \link{mt_example}) with added columns in 
  the data frame specified by \code{save_as} or a new data frame of that name.
}
\description{
Identify (and remove) unusal trajectories based on (a) their most extreme 
points and (b) their similarity to other trajectories. Specifically, the 
function evaluates whether the trajectories most extreme points exceed 
specified limits on the specified dimensions, as well as whether a 
trajectory's average similarity to all other trajectories lies below a
given similarity threshold.
}
\examples{
mt_example <- clean_trajectories(mt_example,
  save_as="data", dimensions = c('xpos','ypos'),
  box_limits = c(2,2), sim_limit  = 2, rescaled_sim = TRUE,
  n_rescale = 20, minkowski_p = 2, clean = FALSE)
  
}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
}

