% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{mt_cluster}
\alias{mt_cluster}
\title{Cluster trajectories.}
\usage{
mt_cluster(data, use = "sp_trajectories", save_as = "clustering",
  dimensions = c("xpos", "ypos"), n_cluster = 5, method = "hclust",
  pointwise = TRUE, minkowski_p = 2, hclust_method = "complete",
  kmeans_nstart = 10)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{save_as}{a character string specifying where the resulting trajectory
data should be stored.}

\item{dimensions}{a character vector specifying the dimensions in the
trajectory array that should be time-normalized. If \code{"all"}, all
trajectory dimensions except the timestamps will be time-normalized.}

\item{method}{character string specifiying the type of clustering procedure. 
Either \link[fastcluster]{hclust} (the default) or \link[stats]{kmeans}.}

\item{pointwise}{boolean specifying the way dissimilarity between the 
trajectories is measured. If \code{TRUE} (the default), \code{mt_distmat} 
measures the average dissimilarity and then sums the results. If 
\code{FALSE}, \code{mt_distmat}  measures dissimilarity once (by treating 
the various points as independent dimensions). Only relevant if
\code{method} is "hclust".}

\item{minkowski_p}{an integer specifying the distance metric. 
\code{minkowski_p = 1} computes the city-block distance, \code{minkowski_p 
= 2} (the default) computes the Euclidian distance, \code{minkowski_p = 3} 
the cubic distance, etc. Only relevant if \code{method} is "hclust".}

\item{hclust_method}{character string specifying the linkage criterion used. 
Passed on to the \code{method} argument of \link[stats]{hclust}. Only 
relevant if \code{method} is "hclust".}

\item{kmeans_nstart}{integer specifying the number of reruns of the kmeans 
procedure. Large numbers minimize the risk of finding local minima.  Passed
on to the \code{nstart} argument of \link[stats]{kmeans}. Only relevant if 
\code{method} is "kmeans".}
}
\value{
A mousetrap data object (see \link{mt_example}) with an additional 
  \link{data.frame} added to it (by default called \code{clustering}) that
  contains the cluster assignments. If a trajectory array was provided
  directly as \code{data}, only the clustering data.frame will be returned.
}
\description{
\code{mt_cluster} performs trajectory clustering.
}
\details{
ToDo: Provide details on different clustering methods and the resulting data.
}
\examples{
# Spatialize trajectories
mt_example <- mt_spatialize(mt_example)
 
# Cluster trajectories
mt_example <- mt_cluster(mt_example, use="sp_trajectories")

}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
  Jonas M. B. Haslbeck  <jonas.haslbeck@gmail.com>
}

