% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{mt_com}
\alias{mt_com}
\title{Detect change of mind trajectories}
\usage{
mt_com(data, use = "trajectories", dimensions = c("xpos", "ypos"),
  save_as = "data", deflection_thresholds = 30, angle_threshold = 50)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import 
functions (see \link{mt_example} for details). Alternatively, a trajectory 
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be 
used.}

\item{dimensions}{a character string specifying which trajectory variables 
should be used. Can be of length 2 or 3 for 2-dimensional or 3-dimensional
assessment respectively.}

\item{save_as}{a character string specifying where the resulting trajectory 
label data should be stored.}

\item{angle_threshold}{an integer specifying the threshold for the trajectory's
angles. Angles are determined for the 2nd, 3rd, ..., nth-1 point by measuring,
for instance, for the 2nd point the angle between the line defined by the 1st
and 2nd point and the line defined by the 2nd and 3rd point.}

\item{deflection_threshold}{an integer vector specifying the thresholds for the
trajectory's deflection. Deflection is measured in terms of movement in the 
direction opposite of the chosen option concerning the first, second, ..., 
nth dimension. Deflection is expressed in terms of percentiles measured from
the midpoint between the options.}
}
\value{
A mousetrap data object (see \link{mt_example}) with added columns in 
  the data frame specified by \code{save_as} or a new data frame of that name.
}
\description{
\code{mt_com} detects change-of-mind (com) trials based on the 
  trajectory's degree of deflection and the development of trajectory
  angles.
}
\examples{
mt_example <- detect_CoM(data=mt_example,
  dimensions = c('xpos','ypos'), save_as="data",
  deflection_thresholds  = 30, angle_threshold = 50)

}
\author{
Dirk U. Wulff
}

