% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectclusters.R
\name{mt_detectclusters}
\alias{mt_detectclusters}
\title{Detect number of clusters}
\usage{
mt_detectclusters(data, use = "sp_trajetcories", dimensions = c("xpos",
  "ypos"), compute = c("stability", "gap", "jump", "slope"), k_seq = 2:15,
  pointwise = TRUE, minkowski_p = 2, method = "hclust",
  hclust_method = "complete", kmeans_nstart = 10, n_bootstrap = 10,
  model_based = FALSE, n_gap = 10, verbose = TRUE)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{dimensions}{a character vector specifying the dimensions in the
trajectory array that should be time-normalized. If \code{"all"}, all
trajectory dimensions except the timestamps will be time-normalized.}

\item{compute}{character vector specifying the to be computed measures. Can
be any subset of \code{c('stability','gap','jump','slope')}.}

\item{k_seq}{ToDo}

\item{pointwise}{boolean specifying the way dissimilarity between the 
trajectories is measured. If \code{TRUE} (the default), \code{mt_distmat} 
measures the average dissimilarity and then sums the results. If 
\code{FALSE}, \code{mt_distmat}  measures dissimilarity once (by treating 
the various points as independent dimensions). Only relevant if
\code{method} is "hclust".}

\item{minkowski_p}{an integer specifying the distance metric. 
\code{minkowski_p = 1} computes the city-block distance, \code{minkowski_p 
= 2} (the default) computes the Euclidian distance, \code{minkowski_p = 3} 
the cubic distance, etc. Only relevant if \code{method} is "hclust".}

\item{method}{character string specifiying the type of clustering procedure
for the stability-based method. Either \code{hclust} or \code{kmeans}.}

\item{hclust_method}{character string specifying the linkage criterion used. 
Passed on to the \code{method} argument of \link[stats]{hclust}. Only 
relevant if \code{method} is "hclust".}

\item{kmeans_nstart}{integer specifying the number of reruns of the kmeans 
procedure. Large numbers minimize the risk of finding local minima.  Passed
on to the \code{nstart} argument of \link[stats]{kmeans}. Only relevant if 
\code{method} is "kmeans".}

\item{n_bootstrap}{ToDo}

\item{model_based}{ToDo}

\item{n_gap}{ToDo}
}
\value{
ToDo.
}
\description{
\code{mt_detectclusters} estimates the number of clusters using various
  methods.
}
\examples{
# Spatialize trajectories
mt_example <- mt_spatialize(mt_example)
 
# Cluster trajectories
mt_example <- mt_detectclusters(mt_example, use="sp_trajectories")

}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
  Jonas M. B. Haslbeck  <jonas.haslbeck@gmail.com>
}

