% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmap.R
\name{mt_diffmap}
\alias{mt_diffmap}
\title{Creates difference-heatmap of two trajectory heatmap images}
\usage{
mt_diffmap(image1, image2, device = "tiff", name = "image",
  directory = getwd(), upscale = 1, smooth_radius = 0,
  colors = c("#00863F", "#FFFFFF", "#FF1900"), n_shades = 10, plot = TRUE)
}
\arguments{
\item{image1}{a data frame containing the trajectory image created by
\code{mt_heatmap}. The first three columns are expected to be named "x",
"y", "img" and contain the respective information.}

\item{image2}{see \code{image1}. Must match the dimensions of \code{image1}.}

\item{device}{a character specifying the plotting device. Either \code{tiff},
\code{png}, or \code{pdf}. Defaults to \code{tiff}.}

\item{name}{a character string specifying the filename}

\item{directory}{a character string specifying the directory for the image
file. Defaults to \code{get.wd()}.}

\item{upscale}{a numeric by which the output resolution of the image is
increased or decreased. Only applies if \code{device} is one of
\code{c("tiff", "png", "pdf")}.}

\item{smooth_radius}{a numeric specifying the standard deviation of the
gaussian smoothing. If zero smoothing is omitted.}

\item{colors}{a character vector specifying the colors used to color
cases where \code{image1 > image2, image1 ≈ image2, image1 < image2},
respectively. Note that the colors will necessarily be used in that
specific order. Defaults to c("#00863F", "#FFFFFF", "#FF1900") which
specifies a green-white-red color gradient.}

\item{n_shades}{integer specifying the number of shades for the color
gradient between the first and second, and the second and third color in
\code{colors}.}

\item{plot}{boolean specifying whether the image should be plotted. If
\code{plot} is \code{FALSE}, then the image will be returned.}
}
\value{
image representing the difference of image1 and image2
}
\description{
\code{mt_diffmap} creates a difference-heatmap of the trajectory
  data using gaussian smoothing.
}
\author{
Dirk U. Wulff
}

