% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmat.R
\name{mt_distmat}
\alias{mt_distmat}
\title{Compute distance matrix.}
\usage{
mt_distmat(data, use = "sp_trajectories", save_as = "distmat",
  dimensions = c("xpos", "ypos"), pointwise = TRUE, minkowski_p = 2)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{save_as}{a character string specifying where the resulting trajectory
data should be stored.}

\item{dimensions}{a character vector specifying which trajectory variables
should be used. Can be of length 2 or 3 for 2-dimensional or 3-dimensional
trajectories respectively.}

\item{pointwise}{boolean specifying the way dissimilarity between the 
trajectories is measured. If \code{TRUE} (the default), \code{mt_distmat}
measures the average dissimilarity and then sums the results. If
\code{FALSE}, \code{mt_distmat}  measures dissimilarity once (by treating
the various points as independent dimensions).}

\item{minkowski_p}{an integer specifying the distance metric. 
\code{minkowski_p = 1} computes the city-block distance, \code{minkowski_p 
= 2} (the default) computes the Euclidian distance, \code{minkowski_p = 3} 
the cubic distance, etc.}
}
\value{
A mousetrap data object (see \link{mt_example}) with an additional 
  object added (by default called \code{distmat}) containing the distance
  matrix. If a trajectory array was provided directly as \code{data}, only
  the distance matrix will be returned.
}
\description{
Computes the point- or vector-wise dissimilarity between
  each pair of trajectories.
}
\examples{
# Spatialize trajectories
mt_example <- mt_spatialize(mt_example)
 
# Compute distance matrix
mt_example <- mt_distmat(mt_example, use="sp_trajectories")

}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
  Jonas M. B. Haslbeck  <jonas.haslbeck@gmail.com>
}

