% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{mt_heatmap}
\alias{mt_heatmap}
\title{Plot trajectory heatmap}
\usage{
mt_heatmap(data, use = "trajectories", dimensions = c("xpos", "ypos"),
  file = "image.tiff", bounds = NULL, xres = 1000, upscale = 1,
  upsample = 1, smooth_radius = 1.5, low_pass = 200,
  auto_enhance = TRUE, mean_intensity = 0.1, plot_dims = TRUE,
  colors = c("black", "blue"), n_shades = c(1000, 10), aggregate_lwd = 0,
  aggregate_col = "black", n_trajectories = 10000, seed = 1,
  verbose = TRUE)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{dimensions}{a character vector specifying the dimensions in the
trajectory array that should be time-normalized. If \code{"all"}, all
trajectory dimensions except the timestamps will be time-normalized.}

\item{file}{a character string giving the name of the file. If \code{NULL}
the R standard device is used for plotting. Otherwise, the plotting 
device is inferred from the file extension. Only supports devices 
\code{tiff()}, \code{png()}, \code{pdf()}.}

\item{bounds}{numeric vector specifying the corners (xmin, ymin, xmax, ymax) of
the plot region. Note that the trajectories' start and end points are mapped to
(0, 0) and (-1, 1.5), respectively. Bounds should therefore include the rectangle
c(-1, 0, 1, 1.5). Defaults to c(-1.4, -.3, 1.4, 1.9).}

\item{upscale}{a numeric value by which the output resolution of the image is increased or
decreased. Only applies if \code{device} is one of \code{c("tiff", "png", "pdf")}.}

\item{upsample}{a numeric by which the number of points used to represent individual
trajectories are increased or decreased. Values of smaller than one will improve speed
but also introduce a certain level of granularity.}

\item{smooth_radius}{a numeric specifying the standard deviation of the gaussian smoothing.
If zero, smoothing is omitted.}

\item{low_pass}{an integer specifying allowed number of counts per pixel. This
arguments limits the maximum pixel color intensity.}

\item{mean_intensity}{a numeric between 0 and 1 specifying the average color intensity
across the entire image. Defaults to .2.}

\item{n_trajectories}{an integer specifying the number of trajectories used to create the
image. If \code{n_trajectories} is smaller than containes in the trajectorie object
specified by \code{use} the \code{n_trajectories} are randomly sampled.}

\item{seed}{an integer specifying the seed used for the trajectory sampling.}

\item{verbose}{logical indicating whether function should report its
progress.}

\item{plot_bounds}{}
}
\description{
\code{mt_heatmap} plots high resolution raw trajectory maps.
}

