% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{mt_heatmap_ggplot}
\alias{mt_heatmap_ggplot}
\title{Plot trajectory heatmap using ggplot.}
\usage{
mt_heatmap_ggplot(...)
}
\arguments{
\item{...}{arguments passed to \link{mt_heatmap}.}
}
\description{
\code{mt_heatmap_ggplot} plots high resolution raw trajectory maps. Note that
this function has beta status.
}
\details{
\code{mt_heatmap_ggplot} is a wrapper for \code{mt_heatmap_plot} to enable. In 
contrast to \code{mt_heatmap_plot} plots created by \code{mt_heatmap_ggplot} 
can be extended using ggplot's \code{+} operator.

For arguments to \code{mt_heatmap_ggplot} see \link{mt_heatmap}
}
\author{
Felix Henninger
Dirk U. Wulff (\email{dirk.wulff@gmail.com})
}
