% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{mt_import_mousetrap}
\alias{mt_import_mousetrap}
\title{Import mouse-tracking data recorded using the mousetrap plug-ins in
OpenSesame.}
\usage{
mt_import_mousetrap(raw_data, xpos_label = "xpos", ypos_label = "ypos",
  timestamps_label = "timestamps", mt_id_label = NULL, split = ", ",
  duplicates = "remove_first", reset_timestamps = TRUE,
  show_progress = TRUE)
}
\arguments{
\item{raw_data}{a data.frame containing the raw data.}

\item{xpos_label}{a character string specifying the name of the column in
which the x-positions are stored (see Details).}

\item{ypos_label}{a character string specifying the name of the column in
which the y-positions are stored (see Details).}

\item{timestamps_label}{a character string specifying the name of the column
in which the timestamps are stored (see Details).}

\item{mt_id_label}{an optional character string specifying the name of the
column that provides a unique ID for every trial in the raw data. If
unspecified, an ID variable will be generated.}

\item{split}{a character string indicating how the different timestamps (and
coordinates) within a trial are separated.}

\item{duplicates}{a character string indicating how duplicate timestamps
within a trial are handled (see Details).}

\item{reset_timestamps}{logical indicating if the first timestamp should be
subtracted from all timestamps within a trial. Default is TRUE as it is
recommended for all following analyses in mousetrap.}

\item{show_progress}{logical indicating whether function should report its
progress.}
}
\value{
A mousetrap data object (see \link{mt_example}).

  If mouse-tracking data were recorded using the mousetrap plug-ins for
  OpenSesame, the unit of the timestamps is milliseconds.
}
\description{
\code{mt_import_mousetrap} accepts a data.frame of (merged) raw data from a
mouse-tracking experiment implemented in OpenSesame using one of the
\href{https://github.com/pascalkieslich/mousetrap-os}{mousetrap plug-ins}.
From this data.frame, \code{mt_import_mousetrap} creates a mousetrap data
object containing the trajectories and additional data for further processing
within the mousetrap package. \code{mt_import_mousetrap} returns a list,
which includes the trajectory data as an array, and all other data as a
data.frame. This data structure can then be passed on to other functions
within this package, such as \link{mt_time_normalize} or
\link{mt_calculate_measures}.
}
\details{
When working with mouse-tracking data that were recorded using the mousetrap
plug-ins for OpenSesame, usually only the \code{raw_data} need to be
provided. All other arguments have sensible defaults.

The column names for the timestamps, x- and y-positions are extracted using
\link{grep} to find the column that starts with the respective character
string (these will typically also contain the name of the item that was used
to record them, such as \code{xpos_track_mouse}). This means that the exact
column names do not have to be provided - as long as mouse-tracking data was
collected only by a single item throughout the experiment (i.e., if only one
column contains timestamps, xpos, and ypos respectively). Otherwise, the
exact column names have to be specified.

\code{duplicates} allows for different options to handle duplicate timestamps
within a trial: \itemize{ \item{\code{remove_first}: First timestamp and
corresponding x-/y-positions are removed (the default).}
\item{\code{remove_last}: Last timestamp and corresponding x-/y-positions are
removed.} \item{\code{ignore}: Duplicates are kept.} }
}
\examples{
mt_example <- mt_import_mousetrap(mt_example_raw)

}
\references{
Mousetrap plug-ins for OpenSesame
  (\url{https://github.com/pascalkieslich/mousetrap-os})
}
\seealso{
\link[readbulk]{read_opensesame} from the \code{readbulk} library
  for reading and combining raw data files that were collected with
  OpenSesame.

\link{mt_import_wide} and \link{mt_import_long} for importing mouse-tracking
data from other sources.
}

