% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mt_map}
\alias{mt_map}
\title{Map trajectories.}
\usage{
mt_map(data, use = "trajectories", dimensions = c("xpos", "ypos"),
  save_as = "map", proto_list = "default", pointwise = TRUE,
  minkowski_p = 2)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{dimensions}{a character vector specifying which trajectory variables
should be used. Can be of length 2 or 3 for 2-dimensional or 3-dimensional
trajectories respectively.}

\item{save_as}{a character string specifying where the resulting trajectory
data should be stored.}

\item{proto_list}{list specifying the list of prototypes the trajectories are
mapped to. Each prototype must be a n x 2 matrix, where the columns correspond
to the to be mapped trajectory dimensions.}

\item{pointwise}{boolean specifying the way dissimilarity between the 
trajectories is measured. If \code{TRUE} (the default), \code{mt_distmat}
measures the average dissimilarity and then sums the results. If
\code{FALSE}, \code{mt_distmat}  measures dissimilarity once (by treating
the various points as independent dimensions).}

\item{minkowski_p}{an integer specifying the distance metric. \code{minkowski_p}
= 1 computes the city-block distance, \code{minkowski_p} = 2 computes the
Euclidian distance, and so on.}

\item{point_wise}{boolean specifying the way dissimilarity between the
trajectories is measured. \code{point-wise = TRUE} measures the average
dissimilarity and then sums the results. \code{point-wise = FALSE} measures
dissimilarity once (by treating the various points as independent dimensions).}

\item{A}{mousetrap data object (see \link{mt_example}) with an additional
object (by default called \code{map}) containing a vector or list indiciating
set of mapped prototypes. If a trajectory array was provided directly as
\code{data}, only the vector or list containing the results will be returned.}
}
\description{
\code{mt_map} maps trajectories onto a predefined set of prototype
  trajectories.
}
\examples{
proto_list <- readRDS(prototypes.RDS)
mt_example <- mt_map(data=mt_example,
  proto_list = proto_list,
  dimensions = c('xpos','ypos'),
  save_as="map")

}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
  Jonas M. B. Haslbeck  <jonas.haslbeck@gmail.com>
}

