% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt_prototypes}
\alias{mt_prototypes}
\title{Mouse trajectory prototypes.}
\format{An object of class \code{list} of length 5.}
\usage{
mt_prototypes
}
\description{
Mouse trajectory prototypes.
}
\keyword{datasets}

