% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{mt_scale_trajectories}
\alias{mt_scale_trajectories}
\title{Standardize trajectories}
\usage{
mt_scale_trajectories(data, use = "trajectories", var_names, prefix = "z_",
  center = TRUE, scale = TRUE)
}
\arguments{
\item{var_names}{character vector giving the labels of the to be 
standardized variables.}

\item{prefix}{character string added to the names of the new 
standardized variables. Must be different from \code{""}.}

\item{center}{logical specifying whether variables should be centered
(i.e., \code{mean = 0}).}

\item{center}{logical specifying whether variables should be scaled
(i.e., \code{sd = 1}).}
}
\value{
A mousetrap data object (see \link{mt_example}) with an additional 
  variable in the object specified by \code{use} containing the standardized
  trajectory variable. If a trajectory array was provided directly as \code{data}, 
  only that array includig now the new variable will be returned.
}
\description{
\code{mt_scale_trajectories} standardizes any trajectory variable to
a mean of 0 (if \code{code = TRUE}) and standard deviation 1 (if \code{
scale = TRUE}).
}
\details{
The new trajectory is added to the object specified in \code{use}.
}
\author{
Dirk U. Wulff (\email{dirk.wulff@gmail.com})
}
