% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mt_smooth}
\alias{mt_smooth}
\title{Smoothing of mousetracking data.}
\usage{
mt_smooth(
  data,
  use = "trajectories",
  save_as = "trajectories_sm",
  dimensions = c("xpos", "ypos"),
  window = gausswindow,
  l = 10,
  symmetric = FALSE,
  normalize = TRUE,
  causal = FALSE,
  ...
)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{save_as}{a character string specifying where the resulting trajectory
data should be stored.}

\item{dimensions}{a character vector specifying the name of the dimension(s)
that should smoothed.}

\item{window}{The windowing function to be used. Should be a function taking
a length parameter and returning a single numeric vector. The entries of
the returned vector will be (after possible normalization) used as weights
for the weighted averaging. By default a non-causal (see caveat below)
Gausswindow is used for filtering.}

\item{l}{length of the averaging window.}

\item{symmetric}{Should the window be made symmetric by using a zero-phase
forward-backward filtering algorithm? This is deactivated by default, to
avoid unexpected non-causal smoothing. Setting this parameter will
also result in an increased window length.}

\item{normalize}{Whether the window should be normalized, i.e. if the weights
should add up to one.}

\item{causal}{Force the filtering to be causal by setting all elements
\eqn{t>l/2} of the window to zero. Requires the window length to
be odd. Also, causal filtering cannot be use for symmetric windows.}

\item{...}{Additional parameters passed to windowing function.}
}
\value{
A mousetrap data object (see \link{mt_example}).

A mousetrap data object (see \link{mt_example}) with an additional
array (by default called \code{trajectories_sm}) that contains the average
trajectory data per dimension interval. If a trajectory array was provided
directly as \code{data}, only the average trajectories will be returned.

If a trajectory array was provided directly as \code{data}, only a named
character vector will be returned.

CAVEAT: If the window contains non-zero elements at indices \eqn{t> l/2},
or if \code{symmetric=TRUE}, then the smoothing is non-causal. I.e.
elements of the trajectories at time \eqn{t} may contain information
about elements at times \eqn{t'>t}. This is also the case if the
default Gausswindow is used.
}
\description{
Smoothing of mousetracking data by weighted averaging using a
windowing function. By default a gausswindow of length 10 is
and width 2.5 is used.
}
\examples{
# smooth the example data
mt_smooth(mt_example)

# smooth the example data using a different length
mt_smooth(mt_example, l=5)

}
\author{
Tillmann Nett
}
