% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialize.R
\name{mt_spatialize}
\alias{mt_spatialize}
\title{Spatialize trajectories}
\usage{
mt_spatialize(data, use = "trajectories", dimensions = c("xpos", "ypos"),
  save_as = "sp_trajectories", n_points = 20, format = "wide")
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{dimensions}{a character string specifying which trajectory variables
should be used. Can be of length 2 or 3 for 2-dimensional or 3-dimensional}

\item{save_as}{a character string specifying where the resulting trajectory
data should be stored.}

\item{n_points}{an integer vector specifying the number of points used to
represent the spatially rescaled trajectories. Length must be 1 or the number
number of trajectories in the data. The values in \code{n_points} may vary.}
}
\value{
A mousetrap data object (see \link{mt_example}) with an additional
  array (by default called \code{spatialized}) containing the
  aligned trajectories. If a trajectory array was provided directly as
  \code{data}, only the aligned trajectories will be returned.
}
\description{
\code{spatialize} re-represents the trajectory spatially so that
  adjacent points on the trajectory become equidistant to each other.
}
\examples{
mt_example <- mt_spatialize(data=mt_example,
  dimensions = c('xpos','ypos'), save_as="rescaled_trajectories",
  n_points = 20, format = 'wide')

}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
  Jonas M. B. Haslbeck  <jonas.haslbeck@gmail.com>
}

