% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{mt_variable_labels}
\alias{mt_variable_labels}
\title{Labels for mousetrap variables.}
\format{An object of class \code{character} of length 6.}
\usage{
mt_variable_labels
}
\description{
A named character vector specifying the names of the variables that mousetrap
uses for labeling the mouse-tracking variables. The default values are 
\code{xpos} for x-positions, \code{ypos} for y-positions, \code{timestamps} 
for timestamps, \code{dist} for distance traveled, \code{vel} for velocity, 
and \code{acc} for acceleration. These labels will be used for labeling the 
respective dimensions in the trajectory array (see \link{mt_example}).
}
\examples{
# the default values
mt_variable_labels <- c(
    timestamps="timestamps",
    xpos="xpos",ypos="ypos",
    dist="dist",vel="vel",acc="acc")

}
\keyword{datasets}

