% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{mt_viewer}
\alias{mt_viewer}
\title{Illustrates individual trajectories}
\usage{
mt_viewer(data, use = "trajectories", dimensions = c("xpos", "ypos"),
  ids = NULL, labels = NULL, sort_variables = NULL, decreasing = NULL,
  overlay = NULL, file = NULL, pagewise = TRUE, nrow = 12, ncol = 6,
  bounds = c(20, 20, 20, 20), pch = 16, cex = 0.2, lwd = NULL,
  plot_box = TRUE, plot_index = TRUE, spatialize = TRUE, rescale = TRUE,
  resolution = 100, start = TRUE, end = TRUE)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{dimensions}{a character vector specifying the dimensions in the
trajectory array that should be time-normalized. If \code{"all"}, all
trajectory dimensions except the timestamps will be time-normalized.}

\item{ids}{a list of character or numeric vectors containing (short)
identifying labels for each trajectory.}

\item{labels}{a list of boolean vectors describing the absence or presence of
a property of interest.}

\item{sort_variables}{a list of character or numeric vectors used to sort the
trajectories, respecting the specific order of sort_variables. If
\code{NULL}, the \code{labels} argument will be used to sort the
trajectories.}

\item{decreasing}{a boolean vector matching the length of either
sort_variables or labels (depending on whether sort_variables is specified)
which specifies for each variable whether the trajectories should be sorted
in an increasing or decreasing order.}

\item{overlay}{a character string specifying the variable used to color the
trajectory.}

\item{file}{a character string specifying the output filename of the
created pdf image. If \code{NULL}, the R standard device will be used. In this
case \code{directory} will be ignored. Otherwise, \code{n_pages} will be
ignored.}

\item{pagewise}{a boolean specifying whether the program should wait for user
input after filling each page of (\code{nrow} times \code{ncol}) plots.}

\item{nrow}{an integer specifying the number of panels per column.}

\item{ncol}{an integer specifying the number of panels per row.}

\item{bounds}{an integer vector of length four specifying the bottom, left,
upper, and right bounds in terms of percent of the distance between the
trajectories' start and end points.}

\item{pch}{plotting character passed on to \link[graphics]{points}.}

\item{cex}{chracter size passed on to \link[graphics]{points}.}

\item{lwd}{line width passed on to \link[graphics]{lines}.}

\item{plot_box}{boolean specifying whether a box should be drawn around each
panel.}

\item{plot_index}{a boolean specifying whether trajectories should be colored
according to the indices of the trajectory points.}

\item{spatialize}{a boolean specifying whether the trajectories should be
spatialized (see \link{mt_spatialize}). Spatialization helps visualize the
trajectory's shape, but loses time information.}

\item{start}{boolean specifying wheter the trajectories' start points should
be aligned.}

\item{end}{boolean specifying wheter the trajectories' end points should be
aligned.}

\item{n_points}{an integer specifying the number of points in the spatialized
trajectory.}
}
\value{
Nothing, if image is plotted to external device, otherwise
  \code{data}.
}
\description{
\code{mt_viewer} plots individual trajectories in a specified order,
including trajectory meta-information, such as whether a trial is identified
as a change-of-mind trial.
}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
}

