% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mt.R
\name{read_mt}
\alias{read_mt}
\title{Read MouseTracker files (.mt)}
\usage{
read_mt(File, columns = "all", add_id = FALSE)
}
\arguments{
\item{File}{a character string specifying the filename of the .mt file.}

\item{add_id}{a boolean specifying whether the filename and a trialnumber should
be added to the data frame.}

\item{variables}{either 'all' or a characteric vector specifying the to be 
extracted variables Defaults to 'all' in which case all existing variables
will be extracted.}
}
\value{
A data frame with one column per trial. Variables are ordered according
  to columns, x-coordinates, y-coordinates, and timestamps.
  array (by default called \code{aligned_trajectories}) containing the
}
\description{
\code{read_mt} reads raw data from MouseTracker files (.mt)
}
\examples{
mt_example <- read_mt(File,columns = 'all', add_id = FALSE)

}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
}

