% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{mt_diffmap}
\alias{mt_diffmap}
\title{Creates a difference-heatmap of two trajectory heatmap images}
\usage{
mt_diffmap(x, y = NULL, cond = NULL, use = "trajectories",
  dimensions = c("xpos", "ypos"), file = "diff_image.tiff", bounds = NULL,
  xres = 500, upscale = 4, smooth_radius = 20, colors = c("#00863F",
  "#000000", "#FF1900"), n_shades = 1000, plot = TRUE, ...,
  verbose = TRUE)
}
\arguments{
\item{x}{an object of class \code{mousetrap}), a trajectory object of class 
\code{array}, or an object of class \code{mt_heatmap_raw} (as created by
\link{mt_heatmap_raw}).}

\item{y}{an object of class \code{mousetrap}), a trajectory object of class 
\code{array}, or an object of class \code{mt_heatmap_raw} (as created by
\link{mt_heatmap_raw}). The class of \code{y} must match the class of 
\code{x}, unless \code{y} is \code{NULL}.}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{dimensions}{a character vector specifying the trajectory variables used
to create the heatmap. The first two entries are used as x and 
y-coordinates, the third, if provided, will be added as color information.}

\item{bounds}{numeric vector specifying the corners (xmin, ymin, xmax, ymax)
of the plot region. For the default \code{bounds = NULL} bounds are
determined based on the data input.}

\item{xres}{an integer specifying the number of pixels along the x-dimension.
An xres of 1000 implies an 1000xN px, where N is determined so that the
trajectories aspect ratio is preserved (provided the \code{bounds} are
unchanged).}

\item{smooth_radius}{a numeric specifying the standard deviation of the
gaussian smoothing. If zero, smoothing is omitted.}

\item{colors}{a character vector specifying the colors used to color
cases of \code{image1 > image2, image1 ~ image2, image1 < image2},
respectively. Note that the colors are used in that specific order.
Defaults to c("#00863F", "#FFFFFF", "#FF1900") which specifies 
a green-black-red color gradient.}

\item{n_shades}{integer specifying the number of shades for the color
gradient between the first and second, and the second and third color in
\code{colors}.}

\item{verbose}{boolean specifying whether progress updates should printed.}
}
\value{
image representing the difference of image1 and image2
}
\description{
\code{mt_diffmap} creates a difference-heatmap of the trajectory
  data using gaussian smoothing.
}
\details{
\code{mt_diffmap} takes two objects that either contain trajectory heatmaps
or from which trajectory heatmaps can be computed. Difference-heatmaps are
constructed analogously to \link{mt_heatmap_raw}.
}
\author{
Dirk U. Wulff (\email{dirk.wulff@gmail.com})
}
