% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{mt_heatmap}
\alias{mt_heatmap}
\title{Plot trajectory heatmap}
\usage{
mt_heatmap(x, use = "trajectories", dimensions = c("xpos", "ypos"),
  filename = "image.tiff", ..., upscale = 1, plot_dims = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x}{usually an object of class mousetrap. Alternatively a trajectory
array or an object of class mt_heatmap_raw.}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{dimensions}{a character vector specifying the dimensions in the
trajectory array that should be time-normalized. If \code{"all"}, all
trajectory dimensions except the timestamps will be time-normalized.}

\item{filename}{a character string giving the name of the file. If \code{NULL} 
the R standard device is used for plotting. Otherwise, the plotting device
is inferred from the file extension. Only supports devices \code{tiff()},
\code{png()}, \code{pdf()}.}

\item{...}{arguments passed to \link{mt_heatmap_raw}.}

\item{upscale}{a numeric value by which the output resolution of the image is
increased or decreased. Only applies if \code{device} is one of
\code{c("tiff", "png", "pdf")}.}

\item{plot_dims}{adds the coordinates of the four image corners to the plot. 
Helps setting \code{bounds}.}

\item{verbose}{logical indicating whether function should report its
progress.}
}
\description{
\code{mt_heatmap} plots high resolution raw trajectory maps.
}
\details{
\code{mt_heatmap} is a wraps \code{mt_heatmap_raw} and provides direct
plotting output in \code{tiff}, \code{png}, or R's default window output. For
further details on how the trajectory heatmaps are constructed see
\link{mt_heatmap_raw}.
}
