% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mt_map}
\alias{mt_map}
\title{Map trajectories to prototypes.}
\usage{
mt_map(data, use = "trajectories", save_as = "prototyping",
  dimensions = c("xpos", "ypos"), prototypes, pointwise = TRUE,
  minkowski_p = 2)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{save_as}{a character string specifying where the resulting data should
be stored.}

\item{dimensions}{a character vector specifying which trajectory variables 
should be used. Can be of length 2 or 3 for two-dimensional or 
three-dimensional trajectories respectively.}

\item{prototypes}{a trajectory array containing the prototypes the 
trajectories are mapped to. As a starting point, the trajectories stored in
\link{mt_prototypes} can be used. See Details and Examples for selecting 
prototypes and creating new ones.}

\item{pointwise}{boolean specifying the way dissimilarity between the 
trajectories is measured (see Details). If \code{TRUE} (the default), 
\code{mt_distmat} measures the average dissimilarity and then sums the 
results. If \code{FALSE}, \code{mt_distmat}  measures dissimilarity once 
(by treating the various points as independent dimensions).}

\item{minkowski_p}{an integer specifying the distance metric. 
\code{minkowski_p = 1} computes the city-block distance, \code{minkowski_p 
= 2} (the default) computes the Euclidian distance, \code{minkowski_p = 3} 
the cubic distance, etc.}
}
\value{
A mousetrap data object (see \link{mt_example}) with an additional 
  \link{data.frame} (by default called \code{prototyping}) that contains the 
  best fitting prototype for each trajectory (the number of the prototype is 
  specified under \code{prototype}, the label of the prototype under 
  \code{prototype_label}) and the distance of the trajectory to the best 
  fitting prototype (\code{min_dist}). If a trajectory array was provided
  directly as \code{data}, only the data.frame containing the results will be
  returned.
}
\description{
\code{mt_map} maps trajectories onto a predefined set of prototype 
trajectories. It first computes distances between the trajectories and each 
of the supplied trajectory types and then assigns each trajectory to the 
prototype that produced the smallest distance.
}
\details{
Mouse trajectories often occur in distinct, qualitative types (see Wulff, 
Haslbeck, Schulte-Mecklenbeck, 2017; Haslbeck, Wulff, Kieslich, Henninger, & 
Schulte-Mecklenbeck, 2017). Common trajectory types are linear trajectories, 
mildly and strongly curved trajctories, and single and multiple 
change-of-mind trials. \code{mt_map} allows to map trajectories to a
predefined set of trajectory types.

First, \code{mt_map} adjusts prototypes to match the coordinate system of the 
trajectories specified by \code{use}. Next, \code{mt_map} computes the 
distances between each trajectory and each of the supplied prototypes (see 
\link{mt_distmat}) and then assigns each trajectory to the prototype that 
produced the smallest distance.

Mapping trajectories to prototypes requires that the endpoints of all
trajectories (and added prototypes) share the same direction, i.e., that all
trajectories end in the top-left corner of the coordinate system
(\link{mt_remap_symmetric} or \link{mt_align} can be used to achieve this).
Furthermore, it is recommended to use spatialized trajectories (see
\link{mt_spatialize}; Haslbeck, Wulff, Kieslich, Henninger, &
Schulte-Mecklenbeck, 2017).
}
\examples{
# Spatialize trajectories
mt_example <- mt_spatialize(mt_example)

# Map trajectories onto standard prototype set
mt_example <- mt_map(mt_example,
  use="sp_trajectories", prototypes=mt_prototypes)
  
# Plot trajectories per cluster
mt_plot(mt_example,use2="prototyping",facet_col="prototype_label")

    
# Map trajectories onto reduced prototype set
mt_example <- mt_map(mt_example,
  use="sp_trajectories",
  prototypes=mt_prototypes[c("straight","curved","cCoM"),,],
  save_as="prototyping_red")


# Map trajectories onto extended prototype set

# Add additional prototypes
mt_prototypes_ext <- mt_add_trajectory(mt_prototypes,
   xpos = c(0,1,-1,1,-1), ypos = c(0,1.5,1.5,1.5,1.5), id = "dCoM3"
)
mt_prototypes_ext <- mt_add_trajectory(mt_prototypes_ext,
   xpos = c(0,0,-1), ypos = c(0,1.5,1.5), id = "neutral"
)

# Map trajectories
mt_example <- mt_map(mt_example,
  use="sp_trajectories", prototypes=mt_prototypes_ext,
  save_as="prototyping_ext")


}
\author{
Dirk U. Wulff (\email{dirk.wulff@gmail.com})

Jonas M. B. Haslbeck (\email{jonas.haslbeck@gmail.com})
}
\references{
Wulff, D. U., Haslbeck, J. M. B., & Schulte-Mecklenbeck, M. 
  (2017). Measuring the (dis-)continuous mind. Manuscript in preparation.
  
  Haslbeck, J. M. B., Wulff, D. U., Kieslich, P. J., Henninger, F., & 
  Schulte-Mecklenbeck, M. (2017). Advanced mouse- and hand-tracking analysis:
  Detecting and visualizing clusters in movement trajectories. Manuscript in 
  preparation.
}

