% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{mt_scale_trajectories}
\alias{mt_scale_trajectories}
\title{Standardize trajectories}
\usage{
mt_scale_trajectories(data, use = "trajectories", var_names, prefix = "z_",
  center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{var_names}{character vector giving the labels of the to be 
standardized variables.}

\item{prefix}{character string added to the names of the new 
standardized variables. Must be different from \code{""}.}

\item{center}{logical specifying whether variables should be centered
(i.e., \code{mean = 0}).}

\item{center}{logical specifying whether variables should be scaled
(i.e., \code{sd = 1}).}
}
\value{
ToDo: add description
}
\description{
\code{mt_scale_trajectories} standardizes any trajectory variable to a mean
of 0 (if \code{code = TRUE}) and standard deviation 1 (if \code{ scale =
TRUE}).The new trajectory is added to the object specified in \code{use}.
}
\examples{
# Normalize velocity
mt_example = mt_derivatives(mt_example)
mt_example = mt_scale_trajectories(mt_example,var_names = 'vel')

}
\author{
Dirk U. Wulff (\email{dirk.wulff@gmail.com})
}
