% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{mt_scale_trajectories}
\alias{mt_scale_trajectories}
\title{Standardize trajectories}
\usage{
mt_scale_trajectories(data, use = "trajectories", var_names,
  transform = NULL, center = TRUE, scale = TRUE, within = FALSE,
  prefix = "z_")
}
\arguments{
\item{data}{a mousetrap data object created using one of the mt_import
functions (see \link{mt_example} for details). Alternatively, a trajectory
array can be provided directly (in this case \code{use} will be ignored).}

\item{use}{a character string specifying which trajectory data should be
used.}

\item{var_names}{character vector giving the labels of the to be 
standardized variables.}

\item{transform}{function that takes a numeric matric as argument and
returns a numeric matric of same size with transformed values. If NULL
the original values are passed on to standardization.}

\item{center}{logical specifying whether variables should be centered
(i.e., \code{mean = 0}). Can be logical vector, in which case the values 
of \code{scale} are mapped to the variables specified in \code{var_names}.}

\item{scale}{logical or numeric specifying the scaling of the variables.
When logical, \code{scale = TRUE} normalizes the trajectory variable
to sd = 1, wheras \code{scale = FALSE} leaves the variable on its original
scale. When numeric, the trajectory variables are scaled by (.i.e., divided
by) the specific value in scale. Can also be numeric vector, in which case
the values of \code{scale} are mapped to the variables specified in
\code{var_names}.}

\item{within}{logical specifying whether trajectory variables should be 
scaled within or across trajectories. I.e., scaling trajectories to mean = 0
and sd = 1 for \code{within == TRUE} means that every trajectory will have
mean = 0 and sd = 1 for with regard to the variables in \code{var_names}. 
\code{within == FALSE} on the other hand renders mean = 0 and sd = 1 only 
true in the aggregate that is when considering all trajectories. Can 
be logical vector, in which case the values of \code{scale} are mapped to 
the variables specified in \code{var_names}.}

\item{prefix}{character string added to the names of the new 
standardized variables. If \code{prefix = ""} the original variables
will be overwritten.}
}
\value{
A mousetrap data object (see \link{mt_example}) with an additional 
  variable in the object specified by \code{use} containing the standardized
  trajectory variable. If a trajectory array was provided directly as \code{data}, 
  only that array includig now the new variable will be returned.
}
\description{
\code{mt_scale_trajectories} standardizes any trajectory variable to
a mean of 0 (if \code{code = TRUE}) and standard deviation 1 (if \code{
scale = TRUE}).
}
\details{
The new trajectory is added to the object specified in \code{use}.
}
\author{
Dirk U. Wulff (\email{dirk.wulff@gmail.com})
}
